+++ test/commands/command_base_test.dart
@@ -25,10 +25,10 @@ class MyCommand extends CommandBase {
         );
 
   // ...........................................................................
-  String? startInputDir;
+  Directory? startInputDir;
 
   @override
-  Future<void> willStart({required String inputDir}) async {
+  Future<void> willStart({required Directory inputDir}) async {
     startInputDir = inputDir;
     await super.willStart(inputDir: inputDir);
   }
@@ -67,24 +67,34 @@ class MyCommand extends CommandBase {
 
 // #############################################################################
 void main() {
+  late CommandRunner<void> runner;
+  late MyCommand myCommand;
+  final messages = <String>[];
+  late List<Directory> repos;
+  late Directory root;
+
+  setUp(() {
+    messages.clear();
+
+    // Create instance
+    myCommand = MyCommand(ggLog: messages.add);
+
+    // Create a command runner
+    runner = CommandRunner<void>(
+      'command-base',
+      'Description goes here.',
+    )..addCommand(myCommand);
+
+    // Create a custom directory structure
+    repos = createSampleRepos();
+    root = repos[0].parent;
+  });
+
   group('CommandBase', () {
     for (final dryRun in ['', '--dry-run', '--no-dry-run']) {
       test('should allow to create custom project processors $dryRun',
           () async {
-        // Create a custom directory structure
-        final repos = createSampleRepos();
-        final root = repos[0].parent;
-        final messages = <String>[];
         final isDryRun = dryRun == '--dry-run' || dryRun == '';
-
-        // Create instance
-        final myCommand = MyCommand(ggLog: messages.add);
-
-        // Create a command runner
-        final runner = CommandRunner<void>(
-          'command-base',
-          'Description goes here.',
-        )..addCommand(myCommand);
 
         // Run the command
         await runner.run(['my-command', '--repos', root.path, dryRun]);
@@ -94,7 +104,7 @@ void main() {
             .sort((a, b) => a.$2.path.compareTo(b.$2.path));
 
         // Check the results
-        expect(myCommand.startInputDir, root.path);
+        expect(myCommand.startInputDir?.path, root.path);
         expect(myCommand.processedProjects.length, repos.length);
         expect(myCommand.didFinishCalled, isTrue);
         expect(myCommand.dryRun, isDryRun);
@@ -111,5 +121,25 @@ void main() {
         expect(true, isNotNull);
       });
     }
+
+    group('special cases', () {
+      test('directory does not contain packages', () async {
+        // Create an empty dir
+        final emptyDir = Directory('${root.path}/empty');
+        await emptyDir.create();
+
+        // Run the command
+        await runner.run([
+          'my-command',
+          '--repos',
+          emptyDir.path,
+        ]);
+
+        // Check the results
+        expect(messages.last, 'No dart packages found in empty.');
+
+        await emptyDir.delete();
+      });
+    });
   });
 }