+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Process dependent packages first
+
 ## [1.0.4] - 2024-04-09
 
 ### Removed
@@ -26,5 +32,6 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_kidney/compare/1.0.4...HEAD
 [1.0.4]: https://github.com/inlavigo/gg_kidney/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/inlavigo/gg_kidney/compare/1.0.2...1.0.3